core: add bookmark module
This bookmark module is the basis for any state that should be stored
in the URL of the application. Any of this data will be stored with
any user-created bookmarks, or if they manually share the URL to an
application.

Bookmarks are purposefully partially opaque. Only the position data is
in cleartext; theu rest is encoded using a very mildly obfuscated
base64, to discourage direct poking.

To use the bookmark system correctly, the part of the application needs
to ensure that it applies state from the bookmark only if there is any
unapplied bookmark fragments, and only writes to the bookmark if there
is no unapplied bookmark data. This ensures no accidental loops occur.
If any data is not committed in due time, the bookmark URL will stop
updating, and the whole bookmark system will jam up.

This commit additionally hooks up the necessary logic to automatically
update the URL fragment leaving no extraenous history.