map: Store current map center and zoom in bookmark
The bookmark, as stored in the store, encodes four values, encoded as
big-endian:
  - Center X (64-bit float)
  - Center Y (64-bit float)
  - Zoom level (32-bit float)
  - Decimal precision (32-bit unsigned integer)

This bookmark data is parsed specially by the BookmarkHelper, ensuring
it is represented in the URL as a pleasing `@x,y,z` value at the start
of the hash.

While zoom level is a floating point, this is not stored as such in the
URL as it would be unnecessarily precise. If the zoom level is less
than one full stop away from the value in the state, it is not
reconfigured.