core: keep track of initially visible map layers
Bookmarks, at this point, encode only the visibility of layers that
have been changed relative to the starting position. (e.g. they encode
which layers a user has enabled as extra, or disabled from the default)
This assumes that an average user does not enable more than a handful
of layers, as each layer takes roughly 2 bytes, or roughly 2.5
characters in the URL. As each layer also takes up considerable amounts
of video memory to render, this is probably a decent compromise.