core: Hook layer visibility up to the bookmark system
The format used for layer visibility in bookmarks is designed to be
relatively compact.

Each layer with changed status (made visible when it originally wasn't
and vice versa) is sorted by layer ID, and encoded into a bytestream
by storing the difference between layer IDs in a varint. This is
followed by a single byte with currently a single bit reserved for
layer visibility.