map: update vector layers while interacting and animating
Any features created with `renderFeatures` currently don't update while
either animating or interacting. These are used for small overlays such
as feature highlighting, and the upcoming geolocation button. For
example, when selecting a feature, moving it off screen, then moving
back without letting go of the cursor, features are only highlighted
once the user lets go of the cursor. This is somewhat inconvenient.